Fix implementation

+++ DTShop.PaymentService.Core/Models/OrderModel.cs
@@ -7,7 +7,7 @@ namespace DTShop.PaymentService.Core.Models
     {
         public int OrderId { get; set; }
         public string Username { get; set; }
-        public int? PaymentId { get; set; }
+        public long? PaymentId { get; set; }
         public string Status { get; set; }
         public decimal TotalCost => OrderItems.Sum(i => i.Amount * i.Item.Price);
         public int TotalAmount => OrderItems.Sum(i => i.Amount);
